+++ tests/data/ex_indentation_tabs.py
@@ -1,0 +1,31 @@
+class Class:
+    def method(self):
+        """
+        Docstring
+                Indent with spaces
+        		Indent with tabs
+        """
+        if True:
+            return None
+
+    def line_length(self):
+        print("xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx")
+        print("xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx")
+
+# output
+
+class Class:
+	def method(self):
+		"""
+		Docstring
+		        Indent with spaces
+				Indent with tabs
+		"""
+		if True:
+			return None
+
+	def line_length(self):
+		print("xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx")
+		print(
+			"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+		)