+++ CHANGES.md
@@ -1,4 +1,17 @@
-## Change Log
+## Change Log (Black EX)
+
+### Unreleased
+
+#### _Black_
+
+- Based on Black [21.5b0](#215b0).
+- Add `--use-tabs` / `-T` to indent with tabs instead of spaces. (#1)
+- Add `--tab-width` / `-w` to specify how many spaces a tab or indentation level
+  represents (default: 4). (#1)
+
+---
+
+## Change Log (Black)
 
 ### 21.5b0